Add basic customisable clipboard shortcuts
